refactor(constellation): unify entity ingestion in services/ingest.ts
